handle broken zip files

+++ optiondata/insert_data.py
@@ -67,24 +67,26 @@ def insert(underlyings, dir):
                                 datafilepath = root + "/" + file
                                 unzippedpath = util.unzip(datafilepath)
                             
-                            df = pd.read_csv(unzippedpath, header=0, dtype={"underlying_symbol": object, "quote_date": object, "root": object, "expiration": object, "strike": np.float64, "option_type": object, "open": np.float64, "high": np.float64, "low": np.float64, "close": np.float64, "trade_volume": np.int64, "bid_size_1545": np.int64, "bid_1545": np.float64, "ask_size_1545": np.int64, "ask_1545": np.float64, "underlying_bid_1545": np.float64, "underlying_ask_1545": np.float64, "bid_size_eod": np.int64, "bid_eod": np.float64, "ask_size_eod": np.int64, "ask_eod": np.float64, "underlying_bid_eod": np.float64, "underlying_ask_eod": np.float64, "vwap": object, "open_interest": np.float64, "delivery_code": object})                                    
-                            
-                            filtered = df[(df['underlying_symbol'] == underlying)]
-                            
-                            if underlying == "^SPX": 
-                                filtered = filtered[(filtered.root != 'BSZ') & (filtered.root != 'BSK') ]  # filter out binary options                    
-                            
-                            filtered['option_type'] = filtered.option_type.str.lower()
-                            filtered['mid_1545'] = (filtered['bid_1545'] + filtered['ask_1545']) / 2 
-                            filtered['underlying_mid_1545'] = (filtered['underlying_bid_1545'] + filtered['underlying_ask_1545']) / 2 
-                            
-                            if len(filtered.index) > 0:
-                                print (str(len(filtered.index)))
-                                dates.add(datestring)
-                                filtered.to_sql(table, engine, if_exists='append', index=False, chunksize=1000)
-                                db.commit()
+                            if unzippedpath is not None: 
+
+                                df = pd.read_csv(unzippedpath, header=0, dtype={"underlying_symbol": object, "quote_date": object, "root": object, "expiration": object, "strike": np.float64, "option_type": object, "open": np.float64, "high": np.float64, "low": np.float64, "close": np.float64, "trade_volume": np.int64, "bid_size_1545": np.int64, "bid_1545": np.float64, "ask_size_1545": np.int64, "ask_1545": np.float64, "underlying_bid_1545": np.float64, "underlying_ask_1545": np.float64, "bid_size_eod": np.int64, "bid_eod": np.float64, "ask_size_eod": np.int64, "ask_eod": np.float64, "underlying_bid_eod": np.float64, "underlying_ask_eod": np.float64, "vwap": object, "open_interest": np.float64, "delivery_code": object})                                    
+                                
+                                filtered = df[(df['underlying_symbol'] == underlying)]
+                                
+                                if underlying == "^SPX": 
+                                    filtered = filtered[(filtered.root != 'BSZ') & (filtered.root != 'BSK') ]  # filter out binary options                    
+                                
+                                filtered['option_type'] = filtered.option_type.str.lower()
+                                filtered['mid_1545'] = (filtered['bid_1545'] + filtered['ask_1545']) / 2 
+                                filtered['underlying_mid_1545'] = (filtered['underlying_bid_1545'] + filtered['underlying_ask_1545']) / 2 
+                                
+                                if len(filtered.index) > 0:
+                                    print (str(len(filtered.index)))
+                                    dates.add(datestring)
+                                    filtered.to_sql(table, engine, if_exists='append', index=False, chunksize=1000)
+                                    db.commit()
                     
-                if unzippedpath != "": 
+                if ((unzippedpath != "") and (unzippedpath is not None)): 
                     os.remove(unzippedpath)
       
     print ("Done inserting data \n")
